refactor(app): split pawwork-sidebar visual subtrees into focused components
Slice 11 of #1056 layout governance line — pure visual extraction.

Extracts four self-contained JSX subtrees out of pawwork-sidebar.tsx into
sibling files so the main component is closer to composition + reactive
wiring only:

- ProjectGroupHeader → pawwork-sidebar-project-group-header.tsx
  (was an in-file helper component; lift to its own file)
- pawwork-side-top (new / search / automations buttons)
  → PawworkSidebarTop in pawwork-sidebar-top.tsx
- pawwork-side-foot (settings button)
  → PawworkSidebarFoot in pawwork-sidebar-foot.tsx
- "All" header + sort dropdown (time / project)
  → PawworkSidebarAllHeader in pawwork-sidebar-all-header.tsx

Verbatim moves: DOM structure, data-component / data-action / data-mode
attributes, aria labels, copy keys, tailwind classes, and Tooltip /
TooltipKeybind / DropdownMenu wrapping all preserved byte-for-byte.
Props flow through unchanged Accessor<T> / callback shape.

pawwork-sidebar.tsx 730 → 551 (-179). Tooltip and TooltipKeybind imports
removed from the parent — both now live only in the extracted children.

## Verification

- bun run typecheck clean
- Full bun run test:unit: 1734 pass / 0 fail
- Targeted sidebar unit tests (interactive invariant, identity,
  session rows, time, sort): 31 pass
- bun run snap sidebar / sidebar-pinned / sidebar-unread: all 3 pass
